Automatically load video from path. Moved videoplayer to custom
